Lower default error severity level from :error to :warn
Each error has a severity level (that takes one of
[:info ,:debug, :warn, :error, :fatal] values) and the default value can be
overwritten both for specific error classes and error instances.

Experience showed that the majority of defined errors are user-side errors
(such as Product ID not Found) and the appropriate level for such errors
would be :warn.

Changing the default level to :warning also means that errors for which the severity level
of :error is justified need to have this declared explicitly, for example:

bad_request class: :sku, severity: :error

Which helps with code clarity and readability.

For this reasons this commit changes the default severity level to :warn.